% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_MOsemiind_app}
\alias{launch_MOsemiind_app}
\title{Launch Shiny App for 'MOsemiind'}
\usage{
launch_MOsemiind_app()
}
\value{
No return value, called for side effects.
}
\description{
This function launches the interactive Shiny application for exploring
copula-based dependence structures in Marshall-Olkin shock models.
}
\examples{
if (interactive()) {
  launch_MOsemiind_app()
}
}
